Redis cache authorization implemented to all classes.

+++ BuildingReport.Business/Concrete/RoleManager.cs
@@ -10,6 +10,8 @@ using BuildingReport.DTO;
 using BuildingReport.DTO.Request;
 using BuildingReport.DTO.Response;
 using BuildingReport.Entities;
+using Microsoft.AspNetCore.Http;
+using StackExchange.Redis;
 using System;
 using System.Collections.Generic;
 using System.Data;
@@ -23,16 +25,33 @@ namespace BuildingReport.Business.Concrete
     {
         private IRoleRepository _roleRepository;
         private readonly IMapper _mapper;
+        private readonly ICacheAuthorityService _cacheAuthorityService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public RoleManager(IMapper mapper)
+        public RoleManager(IMapper mapper, ICacheAuthorityService cacheAuthorityService,IHttpContextAccessor httpContextAccessor)
         {
             _roleRepository = new RoleRepository();
             _mapper = mapper;
+            _cacheAuthorityService = cacheAuthorityService;
+            _httpContextAccessor = httpContextAccessor;
         }
         public RoleResponse CreateRole(RoleRequest roleDTO)
         {
+            // Tokeni headerdan çekiyoruz
+            string token = _httpContextAccessor.HttpContext.Request.Headers["Authorization"].ToString().Substring(7);
+
+
+            //Redis cache'de token keyi ile authorityleri kontrol ediyoruz
+            List<RedisValue> authorityValues = _cacheAuthorityService.CheckCacheAuthority(token);
+
+            // "Create" authority'si var mı kontrol ediyoruz
+            if (!authorityValues.Contains("Create"))
+            {
+                throw new UnauthorizedAccessException();
+            }
+
             _ = roleDTO ?? throw new ArgumentNullException(nameof(roleDTO)," cannot be null.");
-            Role role = _mapper.Map<Role>(roleDTO);
+            Entities.Role role = _mapper.Map<Entities.Role>(roleDTO);
             CheckIfRoleExistsByName(role.Name);
             RoleResponse response = _mapper.Map<RoleResponse>(_roleRepository.CreateRole(role));
             return response;
@@ -40,6 +59,20 @@ namespace BuildingReport.Business.Concrete
 
         public void DeleteRole(long id)
         {
+            // Tokeni headerdan çekiyoruz
+            string token = _httpContextAccessor.HttpContext.Request.Headers["Authorization"].ToString().Substring(7);
+
+
+            //Redis cache'de token keyi ile authorityleri kontrol ediyoruz
+            List<RedisValue> authorityValues = _cacheAuthorityService.CheckCacheAuthority(token);
+
+            // "Delete" authority'si var mı kontrol ediyoruz
+            if (!authorityValues.Contains("Delete"))
+            {
+                throw new UnauthorizedAccessException();
+            }
+
+
             ValidateId(id);
             CheckIfRoleExistsById(id);
             _roleRepository.DeleteRole(id);
@@ -62,8 +95,22 @@ namespace BuildingReport.Business.Concrete
 
         public RoleResponse UpdateRole(UpdateRoleRequest roleDTO)
         {
+            // Tokeni headerdan çekiyoruz
+            string token = _httpContextAccessor.HttpContext.Request.Headers["Authorization"].ToString().Substring(7);
+
+
+            //Redis cache'de token keyi ile authorityleri kontrol ediyoruz
+            List<RedisValue> authorityValues = _cacheAuthorityService.CheckCacheAuthority(token);
+
+            // "Update" authority'si var mı kontrol ediyoruz
+            if (!authorityValues.Contains("Update"))
+            {
+                throw new UnauthorizedAccessException();
+            }
+
+
             CheckIfRoleExistsById(roleDTO.Id);
-            Role role = _mapper.Map<Role>(roleDTO);
+            Entities.Role role = _mapper.Map<Entities.Role>(roleDTO);
             RoleResponse response = _mapper.Map<RoleResponse>(_roleRepository.UpdateRole(role));
             return response;
         }
